Move utility method to container file.
